Fix warning when doing a 64-bit build
- Writing the palette produced warning C4334: result of 32-bit shift implicitly converted to 64 bits.

+++ gifwrite.cpp
@@ -146,7 +146,7 @@ void GIFWriter::WriteHeader(bool loop)
 	// Write (or skip) palette
 	if (lsd.Flags & 0x80)
 	{
-		if (fwrite(GlobalPal, 3, 1 << GlobalPalBits, File) != 1 << GlobalPalBits)
+		if (fwrite(GlobalPal, 3, (size_t)1 << GlobalPalBits, File) != (size_t)1 << GlobalPalBits)
 		{
 			BadWrite();
 			return;
